# api/openapi.yaml
openapi: 3.1.0
info:
  title: Origami FarmOS API
  description: Offline-first, tablet-first farm operating system API for Origami Farms.
    See CONSTITUTION.md and product/MVP_SCOPE.md in the repository root for the product
    principles this API enforces (evidence-based recommendations, worker-observes/manager-decides
    separation, event-sourced history).
  version: 0.1.0
paths:
...
  /api/v1/modules/catalog:
    get:
      tags:
      - employees
      summary: Module Catalog
      description: "Every module the product has, with whether this farm's licence\
        \ for\nit is active (tech spec \xA710 \u2014 modules can be switched off per\
        \ farm's\nsubscription). Any signed-in user may read it; it describes the\n\
        product, not the farm's data."
      operationId: module_catalog_api_v1_modules_catalog_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ModuleCatalogEntry'
                type: array
                title: Response Module Catalog Api V1 Modules Catalog Get
      security:
      - HTTPBearer: []
  /api/v1/me/access:
    get:
      tags:
      - employees
      summary: My Access
      operationId: my_access_api_v1_me_access_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyAccessOut'
      security:
      - HTTPBearer: []
  /api/v1/employees:
    get:
      tags:
      - employees
      summary: List Employees
      operationId: list_employees_api_v1_employees_get
      security:
      - HTTPBearer: []
      parameters:
      - name: include_inactive
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Include Inactive
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmployeeDetailOut'
                title: Response List Employees Api V1 Employees Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - employees
      summary: Create Employee
      operationId: create_employee_api_v1_employees_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDetailOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/employees/{employee_id}:
    get:
      tags:
      - employees
      summary: Get Employee
      operationId: get_employee_api_v1_employees__employee_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: employee_id
        in: path
        required: true
        schema:
          type: string
          title: Employee Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDetailOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - employees
      summary: Update Employee
      operationId: update_employee_api_v1_employees__employee_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: employee_id
        in: path
        required: true
        schema:
          type: string
          title: Employee Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDetailOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - employees
      summary: Deactivate Employee
      description: 'Deactivates rather than deletes: an employee''s name is attached
        to

        every record they ever entered, and the Constitution''s "history is

        never silently deleted" applies to who did the work too.'
      operationId: deactivate_employee_api_v1_employees__employee_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: employee_id
        in: path
        required: true
        schema:
          type: string
          title: Employee Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/employees/{employee_id}/permissions:
    put:
      tags:
      - employees
      summary: Set Employee Permissions
      description: "Replaces the employee's whole responsibility set \u2014 send every\
        \ module\nthey should hold, and anything omitted is revoked."
      operationId: set_employee_permissions_api_v1_employees__employee_id__permissions_put
      security:
      - HTTPBearer: []
      parameters:
      - name: employee_id
        in: path
        required: true
        schema:
          type: string
          title: Employee Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PermissionSet'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDetailOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/farms/me:
    get:
      tags:
      - farms
      summary: My Farm
      description: "The signed-in user's own farm \u2014 just the farm record, for\
        \ the\nSettings screen. `/farms/{farm_id}/bootstrap` below returns this plus\n\
        the whole offline-cache payload, which is far more than a settings\ndisplay\
        \ needs."
      operationId: my_farm_api_v1_farms_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response My Farm Api V1 Farms Me Get
      security:
      - HTTPBearer: []
  /api/v1/farms/{farm_id}/bootstrap:
    get:
      tags:
      - farms
      summary: Bootstrap
      description: "GET /farms/{farm_id}/bootstrap \u2014 initial local cache data\
        \ (tech spec\n\xA712). The tablet writes this straight into SQLite on first\
        \ login /\nwhen activating demo mode, per the offline-first flow in \xA710."
      operationId: bootstrap_api_v1_farms__farm_id__bootstrap_get
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          title: Farm Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BootstrapResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    BootstrapResponse:
      properties:
        farm:
          additionalProperties: true
          type: object
          title: Farm
        users:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Users
        locations:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Locations
        animals:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Animals
        flocks:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Flocks
        fields:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Fields
        inventory_items:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Inventory Items
        tasks:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Tasks
        recommendations:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Recommendations
        server_time:
          type: string
          title: Server Time
        sync_cursor:
          type: string
          title: Sync Cursor
          description: Opaque cursor for the first GET /sync/pull call
      type: object
      required:
      - farm
      - users
      - locations
      - animals
      - flocks
      - fields
      - inventory_items
      - tasks
      - recommendations
      - server_time
      - sync_cursor
      title: BootstrapResponse
      description: "GET /farms/{farm_id}/bootstrap \u2014 initial local-cache payload\n\
        (tech spec \xA712). The tablet stores this verbatim into SQLite on first\n\
        login / demo activation."
...
    EmployeeCreate:
      properties:
        name:
          type: string
          title: Name
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        role:
          type: string
          title: Role
          default: worker
        department:
          anyOf:
          - type: string
          - type: 'null'
          title: Department
        language:
          type: string
          title: Language
          default: en
        job_title:
          anyOf:
          - type: string
          - type: 'null'
          title: Job Title
        employment_status:
          type: string
          title: Employment Status
          default: active
        start_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Start Date
        photo_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Photo Path
        working_days:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Working Days
        working_hours:
          anyOf:
          - type: string
          - type: 'null'
          title: Working Hours
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        password:
          type: string
          title: Password
        permissions:
          items:
            $ref: '#/components/schemas/ModulePermissionIn'
          type: array
          title: Permissions
          default: []
      type: object
      required:
      - name
      - password
      title: EmployeeCreate
    EmployeeDetailOut:
      properties:
        id:
          type: string
          title: Id
        farm_id:
          type: string
          title: Farm Id
        name:
          type: string
          title: Name
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        role:
          type: string
          title: Role
        department:
          anyOf:
          - type: string
          - type: 'null'
          title: Department
        language:
          type: string
          title: Language
        active:
          type: boolean
          title: Active
        job_title:
          anyOf:
          - type: string
          - type: 'null'
          title: Job Title
        employment_status:
          type: string
          title: Employment Status
        start_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Start Date
        photo_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Photo Path
        working_days:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Working Days
        working_hours:
          anyOf:
          - type: string
          - type: 'null'
          title: Working Hours
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        permissions:
          items:
            $ref: '#/components/schemas/ModulePermissionOut'
          type: array
          title: Permissions
          default: []
        full_access:
          type: boolean
          title: Full Access
          default: false
      type: object
      required:
      - id
      - farm_id
      - name
      - role
      - language
      - active
      - employment_status
      title: EmployeeDetailOut
    EmployeeUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        role:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
        department:
          anyOf:
          - type: string
          - type: 'null'
          title: Department
        language:
          anyOf:
          - type: string
          - type: 'null'
          title: Language
        job_title:
          anyOf:
          - type: string
          - type: 'null'
          title: Job Title
        employment_status:
          anyOf:
          - type: string
          - type: 'null'
          title: Employment Status
        start_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Start Date
        photo_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Photo Path
        working_days:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Working Days
        working_hours:
          anyOf:
          - type: string
          - type: 'null'
          title: Working Hours
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Active
        password:
          anyOf:
          - type: string
          - type: 'null'
          title: Password
      type: object
      title: EmployeeUpdate
      description: "Every field optional \u2014 a PATCH only changes what it names."
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ModuleCatalogEntry:
      properties:
        code:
          type: string
          title: Code
        label_en:
          type: string
          title: Label En
        label_ar:
          type: string
          title: Label Ar
        group:
          type: string
          title: Group
        license_code:
          anyOf:
          - type: string
          - type: 'null'
          title: License Code
        licensed_active:
          type: boolean
          title: Licensed Active
          default: true
      type: object
      required:
      - code
      - label_en
      - label_ar
      - group
      title: ModuleCatalogEntry
...
    ModulePermissionIn:
      properties:
        module_code:
          type: string
          title: Module Code
        can_view:
          type: boolean
          title: Can View
          default: true
        can_create:
          type: boolean
          title: Can Create
          default: false
        can_edit:
          type: boolean
          title: Can Edit
          default: false
        can_delete:
          type: boolean
          title: Can Delete
          default: false
        can_approve:
          type: boolean
          title: Can Approve
          default: false
        can_export:
          type: boolean
          title: Can Export
          default: false
        can_assign:
          type: boolean
          title: Can Assign
          default: false
        can_configure:
          type: boolean
          title: Can Configure
          default: false
      type: object
      required:
      - module_code
      title: ModulePermissionIn
    ModulePermissionOut:
      properties:
        module_code:
          type: string
          title: Module Code
        can_view:
          type: boolean
          title: Can View
        can_create:
          type: boolean
          title: Can Create
        can_edit:
          type: boolean
          title: Can Edit
        can_delete:
          type: boolean
          title: Can Delete
        can_approve:
          type: boolean
          title: Can Approve
        can_export:
          type: boolean
          title: Can Export
        can_assign:
          type: boolean
          title: Can Assign
        can_configure:
          type: boolean
          title: Can Configure
      type: object
      required:
      - module_code
      - can_view
      - can_create
      - can_edit
      - can_delete
      - can_approve
      - can_export
      - can_assign
      - can_configure
      title: ModulePermissionOut
...
    MyAccessOut:
      properties:
        user_id:
          type: string
          title: User Id
        role:
          type: string
          title: Role
        full_access:
          type: boolean
          title: Full Access
        modules:
          additionalProperties:
            additionalProperties:
              type: boolean
            type: object
          type: object
          title: Modules
      type: object
      required:
      - user_id
      - role
      - full_access
      - modules
      title: MyAccessOut
      description: "What the signed-in user may do \u2014 the tablet builds its navigation\n\
        and hides its action buttons from this."
...
    PermissionSet:
      properties:
        permissions:
          items:
            $ref: '#/components/schemas/ModulePermissionIn'
          type: array
          title: Permissions
      type: object
      required:
      - permissions
      title: PermissionSet
      description: 'Replaces an employee''s whole responsibility set in one call,
        so the

        tablet''s permission matrix can save exactly what it shows.'
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer